+++ sdk/auth/oauth/testdata/sampleuser.key
