+++ sdk/idp_cert_exchange.go
@@ -8,7 +8,7 @@ import (
 	"github.com/lestrrat-go/jwx/v2/jwk"
 	"github.com/opentdf/platform/lib/ocrypto"
 	"github.com/opentdf/platform/sdk/auth"
-	"github.com/opentdf/platform/sdk/oauth"
+	"github.com/opentdf/platform/sdk/auth/oauth"
 )
 
 type CertExchangeTokenSource struct {
