+++ sdk/idp_token_exchange_token_source.go
@@ -7,7 +7,7 @@ import (
 	"github.com/lestrrat-go/jwx/v2/jwk"
 	"github.com/opentdf/platform/lib/ocrypto"
 	"github.com/opentdf/platform/sdk/auth"
-	"github.com/opentdf/platform/sdk/oauth"
+	"github.com/opentdf/platform/sdk/auth/oauth"
 )
 
 type IDPTokenExchangeTokenSource struct {
